feat(ORCH-024): Deduplicate router helpers into shared module (#92)
* refactor(ORCH-024): deduplicate router helpers into shared module

* WIP [ORCH-024]: auto-committed by router before rework rebase

* fix(ORCH-024): revert unrelated changes and add missing rooms.py import

- Revert alembic/env.py disable_existing_loggers change (scope creep)
- Add missing _helpers import to rooms.py with noqa for unused import (AC3)
- Revert test_error_handling.py to original state (caplog.set_level works correctly)

* fix(ORCH-024): prevent alembic fileConfig from disabling app loggers during tests

Revert caplog.set_level back to caplog.at_level context managers and add
disable_existing_loggers=False to alembic/env.py fileConfig call.
Alembic re-loads env.py per test via upgrade_to_head(), and the default
disable_existing_loggers=True was disabling the media router logger
after the first test created it, causing TestErrorLogging failures.

---------
