+++ alembic/env.py
@@ -12,7 +12,7 @@ from jellyswipe.models.metadata import target_metadata
 config = context.config
 
 if config.config_file_name is not None:
-    fileConfig(config.config_file_name)
+    fileConfig(config.config_file_name, disable_existing_loggers=False)
 
 
 def _resolve_url() -> str:
