+++ jellyswipe/routers/auth.py
@@ -5,7 +5,6 @@ Uses dependency injection for authentication (require_auth) and rate limiting.
 """
 
 import logging
-import traceback
 
 from fastapi import APIRouter, Request, Depends, Response
 from jellyswipe import XSSSafeJSONResponse
@@ -17,39 +16,12 @@ from jellyswipe.dependencies import (
     require_auth,
 )
 from jellyswipe.auth import create_session, destroy_session, resolve_active_room
+from jellyswipe.routers._helpers import make_error_response
 
 _logger = logging.getLogger(__name__)
 
 # Create router with no prefix (D-14)
 auth_router = APIRouter()
-
-
-def make_error_response(
-    message: str, status_code: int, request: Request, extra_fields: dict = None
-) -> XSSSafeJSONResponse:
-    """Create a standardized error response with request ID tracking."""
-    if status_code >= 500:
-        message = "Internal server error"
-    body = {"error": message}
-    body["request_id"] = getattr(request.state, "request_id", "unknown")
-    if extra_fields:
-        body.update(extra_fields)
-    return XSSSafeJSONResponse(content=body, status_code=status_code)
-
-
-def log_exception(exc: Exception, request: Request, context: dict = None) -> None:
-    """Log exception with request context."""
-    log_data = {
-        "request_id": getattr(request.state, "request_id", "unknown"),
-        "route": request.url.path,
-        "method": request.method,
-        "exception_type": type(exc).__name__,
-        "exception_message": str(exc),
-        "stack_trace": traceback.format_exc(),
-    }
-    if context:
-        log_data.update(context)
-    _logger.error("unhandled_exception", extra=log_data)
 
 
 @auth_router.post("/auth/jellyfin-use-server-identity")
